Add files via upload
Fashion-Class-Classification-Using-CNN
The global fashion industry is valued at three trillion dollars and accounts for 2 percent of the world's. GDP the fashion industry is undergoing a dramatic transformation by adopting new computer vision and Machine learning and deep learning techniques.

In this case study we'll look at a hypothetical situation. We assume that if a retailer hired you to build a virtual stylist assistant that looks at customer Instagram and Facebook images and classifies what fashion category they are wearing either bags dresses and pants. The virtual assistant can help the retailer detect and forecast fashion trends and launch targeted marketing campaigns.

In this story we're going to use the fashionmnist data. It's a data set that contains images of bags shoes and dresses. And we're asking the deep network to classify the images into 10 classes.

So we wanted to build kind of an app per se or a model. They can look at images and can tell us exactly what category in this image. Is it like a short. Is it a bag. Is it like a hat. And so on.
That's the whole objective.

The data again they are divided into 28 by 28 greyscale images and the target class is actually No. 1 out of 10 which is kind of a target label which can be categorized as you can see into either like maybe a shoe maybe like like pants. Basically these are the target classes. We have the t shirts trousers pullovers ankle boots sneakers and so on so forth.

Download data set from here https://www.kaggle.com/zalando-research/fashionmnist